fix: Resolve circular dependency between agents and memory
This commit resolves an `ImportError` caused by a circular dependency between `core/ai/agents.py` and `core/ai/memory.py`.

The circular import was created because:
1. `memory.py` imported `AgentMemory` from `agents.py`.
2. `agents.py` and its dependencies eventually led back to importing `memory.py` (via the `__init__.py` files and `server.py`), creating a loop.

The fix involves:
- Creating a new file, `core/ai/schemas.py`, to hold shared data structures.
- Moving the `AgentMemory` dataclass from `agents.py` into `schemas.py`.
- Updating both `agents.py` and `memory.py` to import `AgentMemory` from the new, neutral `schemas.py` file.

This breaks the import cycle and allows the application to initialize correctly.
